+++ core-bundle/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Contao core bundle change log
 
-### DEV
+### 4.4.9 (2017-12-14)
 
  * Show the "invisible" field when editing a form field (see #1199).
  * Only add pages requested via GET to the search index (see #1194).